+++ lib/bloom_remit_client/responses/base_response.rb
@@ -3,10 +3,15 @@ module BloomRemitClient
 
     include Virtus.model
     attribute :raw_response, Object
+    attribute(:body, {
+      Coercers::IndifferentHash,
+      lazy: true,
+      default: :default_body,
+    })
 
     private
 
-    def data
+    def default_body
       json = JSON.parse(raw_response.body)
 
       if json.is_a?(Array)