+++ lib/bloom_remit_client/responses/billers_response.rb
@@ -6,7 +6,7 @@ module BloomRemitClient
     private
 
     def default_billers
-      data.map { |hash| Biller.new(hash) }
+      body.map { |hash| Biller.new(hash) }
     end
 
   end